Update the proposed metadata format
This includes a number of changes -- first, we're introducing the
release keyword; second we're introducing the xmd (extensible metadata)
keyword; third, we're putting all dependencies in the dependencies
block; and fourth, the RPM packages are now SRPM names and you can
specify the repository, the lookaside cache and the sources commit to
build them.

Note this is an incompatible change.

Expect additional changes in the near future.

Signed-off-by: Petr Šabata <contyk@redhat.com>